add: 添加luadb_mock,使得bsp/win32和bsp/linux都支持64bit

+++ bsp/linux/port/luat_gpio_linux.c
@@ -31,7 +31,7 @@ int luat_gpio_setup(luat_gpio_t* gpio) {
     return 0;
 }
 
-int luat_gpio_set(int pin, uint8_t level) {
+int luat_gpio_set(int pin, int level) {
     if (pin < 0 || pin >= LUAT_WIN32_GPIO_COUNT) {
         return -1;
     }
